consistently type lids and lmcs libibnetdisc
Hey Sasha,

The ibnd_port_t lid and lmc are typed uint16_t/uint8_t, but the
ibnd_node_t smalid and smalmc are typed int/int.  This patch turns the
ints into uint16_t/uint8_t.

Naturally, we can do the opposite, but it should be consistent one way
or the other.

Al

--
Albert Chu
chu11@llnl.gov
Computer Scientist
High Performance Systems Division
Lawrence Livermore National Laboratory

From: Albert Chu <chu11@llnl.gov>
Date: Thu, 5 Nov 2009 14:58:11 -0800
Subject: [PATCH] consistently type lids and lmcs libibnetdisc

+++ libibnetdisc/include/infiniband/ibnetdisc.h
@@ -49,8 +49,8 @@ typedef struct ibnd_node {
 	struct ibnd_node *next;	/* all node list in fabric */
 
 	ib_portid_t path_portid;	/* path from "from_node" */
-	int smalid;
-	int smalmc;
+	uint16_t smalid;
+	uint8_t smalmc;
 
 	/* quick cache of switchinfo below */
 	int smaenhsp0;
